+++ inst/vosondash/server/networkGraphsServer.R
@@ -24,6 +24,10 @@ ng_rv <- reactiveValues(
   prune_verts = c()
 )
 
+dt_prev_sel <- reactiveValues(
+  nodes = c()  
+)
+
 # proxy for vertices data table used for row manipulation
 dt_vertices_proxy <- dataTableProxy('dt_vertices')
 
@@ -226,7 +230,8 @@ observeEvent(input$prune_selected_rows_button, {
 })
 
 # add unselected data table rows to pruned vertices list
-observeEvent(input$prune_unselected_rows_button, {
+observeEvent({ input$prune_unselected_rows_button
+               input$nbh_prune_unselected }, {
   pruneListAddOtherNames()
   
   # update prune list select box
@@ -242,6 +247,7 @@ observeEvent(input$prune_unselected_rows_button, {
     prune_list <- temp
   }
   updateSelectInput(session, "pruned_vertices_select", choices = prune_list)
+  # DT::selectRows(dt_vertices_proxy, rownames(graphNodes()))
 })
 
 # remove selected vertices from prune list
@@ -264,7 +270,8 @@ observeEvent(input$prune_return_button, {
 })
 
 # reset prune list
-observeEvent(input$prune_reset_button, {
+observeEvent({ input$prune_reset_button 
+               input$nbh_reset_button }, {
   ng_rv$prune_verts <- c()
   
   updateSelectInput(session, "pruned_vertices_select", choices = character(0))
@@ -274,7 +281,8 @@ observeEvent(input$prune_reset_button, {
 })
 
 # deselect all data table selected rows
-observeEvent(input$prune_deselect_rows_button, { DT::selectRows(dt_vertices_proxy, NULL) })
+observeEvent({ input$prune_deselect_rows_button 
+               input$nbh_deselct_button }, { DT::selectRows(dt_vertices_proxy, NULL) })
 
 # nodes clicked event in visnetwork
 observeEvent(input$vis_node_select, {
@@ -290,6 +298,52 @@ observeEvent(input$vis_node_select, {
   sel <- which(rownames(dt_vertices) %in% sel) # require indices not row names
   
   DT::selectRows(dt_vertices_proxy, sel)
+})
+
+# observeEvent(input$vis_nbh_node_select, {
+#   g <- graphFilters()
+#   dt_vertices <- graphNodes()
+#   plot_sel_nodes <- row.names(dt_vertices)[dt_vertices$name %in% input$vis_nbh_node_select]
+#   DT::selectRows(dt_vertices_proxy, which(rownames(dt_vertices) %in% plot_sel_nodes))
+#   sel_rows <- row.names(dt_vertices)[c(input$dt_vertices_rows_selected)]
+#   dt_prev_sel$nodes <- sel_rows
+#   shinyjs::enable("nbh_undo_button")
+#   sel_row_names <- V(g)[V(g)$id %in% sel_rows]$name
+#   order <- 1
+#   g_ego <- make_ego_graph(g, order = order, nodes = sel_row_names, mode = "all", mindist = 0)
+#   ids <- unlist(sapply(g_ego, function(x) V(x)$id))
+#   sel <- which(rownames(dt_vertices) %in% ids)
+#   DT::selectRows(dt_vertices_proxy, sel)  
+# })
+
+observeEvent(input$nbh_select_button, {
+  if (length(input$dt_vertices_rows_selected) < 1) { return() }
+  g <- graphFilters()
+  dt_vertices <- graphNodes()
+  sel_rows <- row.names(dt_vertices)[c(input$dt_vertices_rows_selected)]
+  
+  dt_prev_sel$nodes <- sel_rows
+  shinyjs::enable("nbh_undo_button")
+  
+  sel_row_names <- V(g)[V(g)$id %in% sel_rows]$name
+  
+  order <- input$nbh_order_select
+  g_ego <- make_ego_graph(g, order = order, nodes = sel_row_names, mode = "all", mindist = 0)
+  ids <- unlist(sapply(g_ego, function(x) V(x)$id))
+  sel <- which(rownames(dt_vertices) %in% ids)
+  
+  DT::selectRows(dt_vertices_proxy, sel)
+})
+
+observeEvent(input$nbh_undo_button, {
+  if (length(dt_prev_sel$nodes) > 0) {
+    DT::selectRows(dt_vertices_proxy, NULL)
+    sel <- which(rownames(graphNodes()) %in% dt_prev_sel$nodes)
+    DT::selectRows(dt_vertices_proxy, sel)
+    
+    dt_prev_sel$nodes <- c()
+    shinyjs::disable("nbh_undo_button")
+  }
 })
 
 # reset node size slider when changed to none
@@ -548,6 +602,9 @@ setGraphFilterControls <- reactive({
     shinyjs::enable("graph_component_slider")
 
     updateComponentSlider(g, isolate(input$graph_component_type_select))
+    
+    dt_prev_sel$nodes <- c()
+    shinyjs::reset("nbh_undo_button")
     
     # update the categorical attribute select box
     if (!is.null(ng_rv$graph_cats) && length(ng_rv$graph_cats) > 0) {