+++ R/vosonSML.R
@@ -11,19 +11,6 @@ getVosonSMLVersion <- function() {
   "unknown"
 }
 
-#' @title Return logical if vosonSML version later than 0.29
-#' 
-#' @description This function returns if the installed version of vosonSML is later than v0.29
-#'
-#' @return Logical.
-#' 
-#' @keywords internal
-#' @export
-isVosonSML0290 <- function() {
-  if (utils::packageVersion("vosonSML") >= "0.29.0") { return(TRUE) }
-  FALSE
-}
-
 #' @title Create an auth token with twitter app dev keys
 #' 
 #' @description This function is a wrapper for \code{vosonSML::Authenticate} with twitter app developer keys. The